Rearrange .tools content into separated files
