+++ alfred/marshmallow_utils/fields/password.py
@@ -1,6 +1,6 @@
 from marshmallow import ValidationError, fields
 
-from alfred.tools import only_digits
+from alfred.tools.core import only_digits
 
 
 class PasswordNumberField(fields.String):
