+++ alfred/sqs/exceptions.py
@@ -1,2 +1,6 @@
+class SQSTaskError(Exception):
+    pass
+
+
 class SQSTaskMaxRetriesExceededError(Exception):
     pass
