--- alfred/tools/__init__.py
@@ -1,8 +1,0 @@
-from .core import has_accent, is_br_phone_valid, is_email_valid, only_digits
-
-__all__ = [
-    "has_accent",
-    "is_br_phone_valid",
-    "is_email_valid",
-    "only_digits",
-]
